+++ editor/editor_node.h
@@ -65,6 +65,7 @@ class TabContainer;
 class TextureRect;
 class TextureProgressBar;
 class Tree;
+class VBoxContainer;
 class VSplitContainer;
 class Window;
 
@@ -269,7 +270,6 @@ private:
 
 	EditorCommandPalette *command_palette = nullptr;
 	EditorExport *editor_export = nullptr;
-	EditorInterface *editor_interface = nullptr;
 	EditorLog *log = nullptr;
 	EditorNativeShaderSourceVisualizer *native_shader_source_visualizer = nullptr;
 	EditorPlugin *editor_plugin_screen = nullptr;
