+++ editor/editor_run_script.h
@@ -31,9 +31,11 @@
 #ifndef EDITOR_RUN_SCRIPT_H
 #define EDITOR_RUN_SCRIPT_H
 
+#include "core/object/gdvirtual.gen.inc"
 #include "core/object/ref_counted.h"
-#include "editor/editor_plugin.h"
+#include "core/object/script_language.h"
 
+class EditorInterface;
 class EditorNode;
 
 class EditorScript : public RefCounted {
