+++ editor/plugins/gpu_particles_collision_sdf_editor_plugin.cpp
@@ -30,6 +30,7 @@
 
 #include "gpu_particles_collision_sdf_editor_plugin.h"
 
+#include "editor/editor_interface.h"
 #include "editor/editor_node.h"
 #include "editor/gui/editor_file_dialog.h"
 
@@ -193,7 +194,7 @@ GPUParticlesCollisionSDF3DEditorPlugin::GPUParticlesCollisionSDF3DEditorPlugin()
 	probe_file->set_file_mode(EditorFileDialog::FILE_MODE_SAVE_FILE);
 	probe_file->add_filter("*.exr");
 	probe_file->connect("file_selected", callable_mp(this, &GPUParticlesCollisionSDF3DEditorPlugin::_sdf_save_path_and_bake));
-	get_editor_interface()->get_base_control()->add_child(probe_file);
+	EditorInterface::get_singleton()->get_base_control()->add_child(probe_file);
 	probe_file->set_title(TTR("Select path for SDF Texture"));
 
 	GPUParticlesCollisionSDF3D::bake_begin_function = bake_func_begin;
