+++ editor/register_editor_types.cpp
@@ -35,6 +35,7 @@
 #include "editor/editor_command_palette.h"
 #include "editor/editor_feature_profile.h"
 #include "editor/editor_file_system.h"
+#include "editor/editor_interface.h"
 #include "editor/editor_node.h"
 #include "editor/editor_paths.h"
 #include "editor/editor_resource_picker.h"
@@ -230,10 +231,14 @@ void register_editor_types() {
 
 	GLOBAL_DEF("editor/version_control/plugin_name", "");
 	GLOBAL_DEF("editor/version_control/autoload_on_startup", false);
+
+	EditorInterface::create();
 }
 
 void unregister_editor_types() {
 	EditorNode::cleanup();
+	EditorInterface::free();
+
 	if (EditorPaths::get_singleton()) {
 		EditorPaths::free();
 	}
